refactor: extract reusable building blocks from K8s and Nacos discovery
Extract kubernetes/core.lua and nacos/client.lua as composable building
blocks from the existing init.lua modules.

kubernetes/core.lua provides:
- Config parsing (get_apiserver, setup_namespace_selector, etc.)
- Parameterized endpoint callbacks via create_endpoint_callbacks()
  with key_prefix for multi-registry dict isolation and
  duplicate_port_number for flexible port key storage
- Handle factory (create_handle) and lifecycle (start_fetch)
- Node resolution with LRU cache (resolve_nodes)
- Dict helpers (dump_endpoints_from_dict)

nacos/client.lua provides:
- Stateless HTTP client primitives (request, get_token_param)
- HMAC-SHA1 signing (get_signed_param)
- URL building (build_base_uri)
- Instance fetching (fetch_from_host) with configurable key_builder,
  preserve_metadata, and timeout
- Service scanning (get_nacos_services) with optional filter

informer_factory.lua enhancements:
- Graceful stop support (informer.stop flag) checked in watch() and
  list_watch() loops
- Configurable ssl_verify (from apiserver config, default false)
- ssl_server_name support for HTTPS connections

schema.lua:
- Add ssl_verify option to service config for both single and
  multiple mode

Both init.lua modules are refactored to use the new building blocks
with no behavior change for existing users.
